RUM-9958: Align datadog initialization with ios for benchmark app
