+++ sample/benchmark/src/main/java/com/datadog/benchmark/sample/di/app/BenchmarkAppComponent.kt
@@ -14,8 +14,7 @@ import javax.inject.Singleton
 
 @Component(
     modules = [
-        AppModule::class,
-        DatadogModule::class
+        AppModule::class
     ]
 )
 @Singleton
